feat: add AppLifecycleUI class
